+++ src/components/Autocomplete/Suggestion/index.js
@@ -1,0 +1,2 @@
+import Suggestion from './Suggestion';
+export default Suggestion;