+++ src/components/Autocomplete/examples/index.js
@@ -1,0 +1,14 @@
+/* eslint-disable react/jsx-key */
+import React from 'react';
+import Autocomplete from '../index';
+
+export default [
+  {
+    examples: [
+      <Autocomplete
+        placeholder='typehead...'
+        suggestions={ ['foo', 'bar'] }
+      />,
+    ],
+  },
+];